:white_check_mark: Move interface-related tests to their own folder
